services/vision/pyproject.toml
[project]
name = "qaongdur-vision"
version = "0.1.0"
description = "Qaongdur vision pipeline scaffold and demo ingestion service."
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.1.0",
  "opencv-python-headless>=4.10.0",
  "Pillow>=11.1.0",
  "fastapi>=0.115.0",
  "pydantic-settings>=2.7.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]
models = [
  "open-clip-torch>=2.30.0",
  "ultralytics>=8.3.0",
]

[project.scripts]
qaongdur-vision = "vision_service.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vision_service"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
